Split the onboarding and QR code
